refactor: toServerDateFormat 함수 재사용성을 위해 utils 패키지에 리펙토링 https://github.com/Team-PATCH/Harmony-Android/pull/48#discussion_r1888580235
