--- core/data/src/main/java/com/teampatch/core/data/Utils.kt
@@ -1,5 +1,0 @@
-package com.teampatch.core.data
-
-import java.time.format.DateTimeFormatter
-
-internal val serverDateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
