+++ core/data/src/main/java/com/teampatch/core/data/repository/AuthenticationRepositoryImpl.kt
@@ -1,15 +1,12 @@
 package com.teampatch.core.data.repository
 
 import com.agvber.core.authentication.kakao.KakaoLoginService
-import com.teampatch.core.data.serverDateTimeFormatter
+import com.teampatch.core.data.utils.toServerDateFormat
 import com.teampatch.core.domain.entity.TokenManager
 import com.teampatch.core.domain.model.LoginResult
 import com.teampatch.core.domain.repository.AuthenticationRepository
 import com.teampatch.core.network.UserRemoteDataSource
 import com.teampatch.core.network.model.user.SignupOrLoginRequestBody
-import java.time.LocalDateTime
-import java.time.ZoneId
-import java.util.Date
 import javax.inject.Inject
 
 class AuthenticationRepositoryImpl @Inject constructor(
@@ -34,13 +31,5 @@ class AuthenticationRepositoryImpl @Inject constructor(
         tokenManager.setAccessToken(signupOrLoginResponse.token)
 
         return LoginResult(groupId = signupOrLoginResponse.user.groupId)
-    }
-
-    private fun Date.toServerDateFormat(): String {
-        return LocalDateTime.ofInstant(
-            toInstant(),
-            ZoneId.systemDefault()
-        )
-            .format(serverDateTimeFormatter)
     }
 }
